pyproject.toml
[tool.poetry]
name = "test-tonga-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
airtable-python-wrapper = "^0.15.1"
APScheduler = "^3.7.0"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

# TODO: add flask-mail